Online Resume Builder
Developed using Android, this Mobile Application is for creating resume in few minutes by selecting template and entering the data. This application generates
PDF file and save to mobile or it allows to update and edit the resume using SQLite. It also allows the user to create resume using 3 different languages.